Solution to change the permission of the file
Question:
Change the permission of the file /user/acadgild/hadoop/max-temp.txt, such that only the owner and the group members have full control over the file. Others do not have any control over it.
Answer:
for this,  I used the chmod command (is used to change the permissions of files or directories). The required permission for the file is obtained typing "chmod 770 max-temp.txt" in shell